Use update count to determine whether password reset token is valid
This also avoids the odd situation of `checktoken` returning a positive for tokens that will turn out to be invalid.
